backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-generator==1.10
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
distro==1.9.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.111.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Flask==3.1.2
flask-cors==6.0.2
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.114.0
gunicorn==21.2.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.26.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
MouseInfo==0.1.3
mypy_extensions==1.1.0
npx==0.1.6
numpy==2.3.5
oauthlib==3.3.1
openai==2.16.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.0
passlib==1.7.4
php==1.2.1
pillow==12.1.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0

PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyparsing==3.3.2
pypdf==4.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pyswisseph==2.10.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.22
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.3
razorpay==2.0.0
reportlab==4.0.9
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.40.0
setuptools==80.10.2
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45

tqdm==4.67.2
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
undetected-chromedriver==3.5.5
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.27.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.5
wsproto==1.3.2
zipp==3.23.0
